Elliptical trig approximation function
Fixed up the elliptical trig function, actually restricting the range to
[0, K / 4). Still some stair-stepping, unsure of the exact cause, adding
more Maclaurin terms didn't help a lot, but if we pre-calculate and
store the coefficients (since k (or m) changes rarely) we can afford to
throw in quite a few terms. Will continue to read to see if there are
other error mitigation strategies.